Convert React examples to .compatability.stories
